tests/wpt/tests/css/css-position/parsing/right-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing right with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-right">
<meta name="assert" content="right supports the full grammar 'auto | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("right", "auto");

test_valid_value("right", "-10px");
test_valid_value("right", "-20%");
test_valid_value("right", "calc(2em + 3ex)");
</script>
</body>
</html>
